+++ resources/config/sluggable.php
@@ -138,4 +138,10 @@ return [
     
     'onUpdate' => false,
 
+    /**
+     * If the default slug engine of cocur/slugify is used, this array of
+     * configuration options will be used when instantiating the engine.
+     */
+    'slugEngineOptions' => [],
+
 ];