+++ tests/BaseTests.php
@@ -7,6 +7,7 @@ use Cviebrock\EloquentSluggable\Tests\Models\PostShortConfig;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomCallableMethod;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomEngine;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomEngine2;
+use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomEngineOptions;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomMethod;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomSeparator;
 use Cviebrock\EloquentSluggable\Tests\Models\PostWithCustomSource;
@@ -203,10 +204,9 @@ class BaseTests extends TestCase
      */
     public function testNonSluggableModels()
     {
-        $post = new PostNotSluggable([
+        $post = PostNotSluggable::create([
             'title' => 'My First Post'
         ]);
-        $post->save();
         $this->assertEquals(null, $post->slug);
     }
 
@@ -368,10 +368,9 @@ class BaseTests extends TestCase
      */
     public function testCustomEngineRules()
     {
-        $post = new PostWithCustomEngine([
+        $post = PostWithCustomEngine::create([
             'title' => 'The quick brown fox jumps over the lazy dog'
         ]);
-        $post->save();
         $this->assertEquals('tha-qaack-brawn-fax-jamps-avar-tha-lazy-dag', $post->slug);
     }
 
@@ -380,11 +379,18 @@ class BaseTests extends TestCase
      */
     public function testCustomEngineRules2()
     {
-        $post = new PostWithCustomEngine2([
+        $post = PostWithCustomEngine2::create([
             'title' => 'The quick brown fox/jumps over/the lazy dog'
         ]);
-        $post->save();
         $this->assertEquals('the-quick-brown-fox/jumps-over/the-lazy-dog', $post->slug);
+    }
+
+    public function testCustomEngineOptions()
+    {
+        $post = PostWithCustomEngineOptions::create([
+            'title' => 'My First Post'
+        ]);
+        $this->assertEquals('My-First-Post', $post->slug);
     }
 
     /**
@@ -405,10 +411,9 @@ class BaseTests extends TestCase
      */
     public function testEmptySeparator()
     {
-        $post = new PostWithEmptySeparator([
+        $post = PostWithEmptySeparator::create([
             'title' => 'My Test Post'
         ]);
-        $post->save();
         $this->assertEquals('mytestpost', $post->slug);
     }
 
@@ -417,11 +422,10 @@ class BaseTests extends TestCase
      */
     public function testMultipleSlugs()
     {
-        $post = new PostWithMultipleSlugs([
+        $post = PostWithMultipleSlugs::create([
             'title' => 'My Test Post',
             'subtitle' => 'My Subtitle',
         ]);
-        $post->save();
 
         $this->assertEquals('my-test-post', $post->slug);
         $this->assertEquals('my.subtitle', $post->dummy);
@@ -432,10 +436,9 @@ class BaseTests extends TestCase
      */
     public function testSubscriptCharacters()
     {
-        $post = new Post([
+        $post = Post::create([
             'title' => 'RDA-125-15/30/45m³/h CAV'
         ]);
-        $post->save();
 
         $this->assertEquals('rda-125-15-30-45m3-h-cav', $post->slug);
     }