+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Change Log
 
-## x.x.x - xx-xxx-xxxx
+## 8.0.3 - 15-Jan-2021
 
+- change package to slug models on Eloquent's `saved` event, rather than
+  `saving`
+  - this adds a few more SQL queries per save, but allows for the use of
+    the primary key field in the `source` configuration (see #539 and #448)
+- add `slugEngineOptions` configuration option (see #454, thanks @Adioz01)
 - move automated testing from travis-ci to Github actions (#534, thanks @cbl)
 - clean up some third-party tools and badges